## refactor(services): standardize service interface and implementation imports
Improves the organization, maintainability, and consistency of our service layer by centralizing type imports and standardizing implementation file naming conventions across the codebase.

```
apps/app/src/lib/services/
├── completion/
│   ├── _types.ts
│   ├── anthropic.ts
│   ├── google.ts
│   ├── groq.ts
│   └── openai.ts
```

- **Centralized Service Types:**
  All service-related types are now imported/exported from a dedicated `_types` module within each service folder.

- **Consistent Implementation Naming:**
  Service implementation files are now named according to the `[platform].ts` convention (e.g., `web.ts`, `desktop.ts`).

Centralized types and consistent naming make future refactors and platform-specific changes less error-prone.  The codebase now follows a single, predictable pattern for service imports and implementations, improving code quality and maintainability.
