+++ apps/app/src/lib/services/clipboard/_types.ts
