+++ apps/app/src/lib/services/completion/_types.ts
