+++ apps/app/src/lib/services/completion/groq.ts
@@ -1,7 +1,7 @@
 import { Err, Ok } from '@epicenterhq/result';
 import { z } from 'zod';
-import type { HttpService } from '../http/HttpService';
-import type { CompletionService } from './types';
+import type { HttpService } from '$lib/services/http/_types';
+import type { CompletionService } from './_types';
 
 export function createGroqCompletionService({
 	apiKey,
