+++ apps/app/src/lib/services/download/web.ts
@@ -1,5 +1,5 @@
 import { tryAsync } from '@epicenterhq/result';
-import type { DownloadService, DownloadServiceError } from './DownloadService';
+import type { DownloadService, DownloadServiceError } from './_types';
 
 export function createDownloadServiceWeb(): DownloadService {
 	return {
