+++ apps/app/src/lib/services/transcription/whisper/elevenlabs.ts
@@ -1,7 +1,7 @@
 import { Err, Ok } from '@epicenterhq/result';
 import type { WhisperingError } from '@repo/shared';
 import { ElevenLabsClient } from 'elevenlabs';
-import type { TranscriptionService } from './TranscriptionService';
+import type { TranscriptionService } from '../_types';
 
 type ModelName = 'scribe_v1';
